feat: use src/ dir for holding files
