+++ src/pages/[...path].tsx
@@ -1,14 +1,17 @@
 import Head from 'next/head'
+import { useRouter } from 'next/router'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import siteConfig from '../config/site.config'
+import siteConfig from '../../config/site.config'
 import Navbar from '../components/Navbar'
 import FileListing from '../components/FileListing'
 import Footer from '../components/Footer'
 import Breadcrumb from '../components/Breadcrumb'
 import SwitchLayout from '../components/SwitchLayout'
 
-export default function Home() {
+export default function Folders() {
+  const { query } = useRouter()
+
   return (
     <div className="flex min-h-screen flex-col items-center justify-center bg-white dark:bg-gray-900">
       <Head>
@@ -17,12 +20,12 @@ export default function Home() {
 
       <main className="flex w-full flex-1 flex-col bg-gray-50 dark:bg-gray-800">
         <Navbar />
-        <div className="mx-auto w-full max-w-5xl sm:p-4 py-4">
-          <nav className="mb-4 flex items-center justify-between sm:pl-1 sm:px-0 px-4">
-            <Breadcrumb />
+        <div className="mx-auto w-full max-w-5xl py-4 sm:p-4">
+          <nav className="mb-4 flex items-center justify-between space-x-3 px-4 sm:px-0 sm:pl-1">
+            <Breadcrumb query={query} />
             <SwitchLayout />
           </nav>
-          <FileListing />
+          <FileListing query={query} />
         </div>
       </main>
