+++ src/pages/_document.tsx
@@ -1,5 +1,5 @@
 import Document, { Head, Html, Main, NextScript } from 'next/document'
-import siteConfig from '../config/site.config'
+import siteConfig from '../../config/site.config'
 
 class MyDocument extends Document {
   render() {
