+++ src/pages/api/index.ts
@@ -3,8 +3,8 @@ import { posix as pathPosix } from 'path'
 import type { NextApiRequest, NextApiResponse } from 'next'
 import axios from 'axios'
 
-import apiConfig from '../../config/api.config'
-import siteConfig from '../../config/site.config'
+import apiConfig from '../../../config/api.config'
+import siteConfig from '../../../config/site.config'
 import { revealObfuscatedToken } from '../../utils/oAuthHandler'
 import { compareHashedToken } from '../../utils/protectedRouteHandler'
 import { getOdAuthTokens, storeOdAuthTokens } from '../../utils/odAuthTokenStore'
