+++ src/pages/api/name/[name].ts
