+++ src/pages/index.tsx
@@ -1,17 +1,14 @@
 import Head from 'next/head'
-import { useRouter } from 'next/router'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import siteConfig from '../config/site.config'
+import siteConfig from '../../config/site.config'
 import Navbar from '../components/Navbar'
 import FileListing from '../components/FileListing'
 import Footer from '../components/Footer'
 import Breadcrumb from '../components/Breadcrumb'
 import SwitchLayout from '../components/SwitchLayout'
 
-export default function Folders() {
-  const { query } = useRouter()
-
+export default function Home() {
   return (
     <div className="flex min-h-screen flex-col items-center justify-center bg-white dark:bg-gray-900">
       <Head>
@@ -20,12 +17,12 @@ export default function Folders() {
 
       <main className="flex w-full flex-1 flex-col bg-gray-50 dark:bg-gray-800">
         <Navbar />
-        <div className="mx-auto w-full max-w-5xl sm:p-4 py-4">
-          <nav className="mb-4 flex items-center justify-between space-x-3 sm:pl-1 sm:px-0 px-4">
-            <Breadcrumb query={query} />
+        <div className="mx-auto w-full max-w-5xl py-4 sm:p-4">
+          <nav className="mb-4 flex items-center justify-between px-4 sm:px-0 sm:pl-1">
+            <Breadcrumb />
             <SwitchLayout />
           </nav>
-          <FileListing query={query} />
+          <FileListing />
         </div>
       </main>
