+++ src/pages/onedrive-vercel-index-oauth/step-1.tsx
@@ -4,8 +4,8 @@ import { useRouter } from 'next/router'
 import { useTranslation, Trans } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import siteConfig from '../../config/site.config'
-import apiConfig from '../../config/api.config'
+import siteConfig from '../../../config/site.config'
+import apiConfig from '../../../config/api.config'
 import Navbar from '../../components/Navbar'
 import Footer from '../../components/Footer'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
