+++ src/utils/getPreviewType.ts
